+++ CHANGELOG.md
@@ -23,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improve `Throttling` adoptor
   - Freeze when getting `RetryAfter(_)` error
   - Retry requests that previously returned `RetryAfter(_)` error
-- `RequestError::RetryAfter` now has a `u32` field instead of `i32`
+- `RequestError::RetryAfter` now has a `Duration` field instead of `i32`
 
 ### Added